Work on public interface
