+++ test/api_tests/CMakeLists.txt
@@ -4,6 +4,7 @@ add_executable(api_tests
     helpers.hpp
     matchers.hpp
 
+    tiro_functions_test.cpp
     tiro_objects_test.cpp
     tiro_vm_test.cpp
